requirements.txt
pandas>=1.3
python-docx>=0.8.11